requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
gast==0.3.3
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
pkg-resources==0.0.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scapy==2.4.3
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
tb-nightly==2.3.0a20200709
tensorboard==2.0.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020070901
tf-nightly==2.4.0.dev20200709
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
